+++ include/alc_max_m10s_sensor.h
@@ -158,7 +158,7 @@ public:
    * @param bus_manager Reference to the I2C bus manager.
    * @param address I2C address of the sensor (default: 0x42).
    */
-  explicit MaxM10sSensor(I2CBusManager& bus_manager, uint8_t address = 0x42);
+  explicit MaxM10sSensor(I2CBusManager& bus_manager, uint16_t address = 0x42);
   ~MaxM10sSensor();
 
   MaxM10sSensor() = delete;
@@ -245,7 +245,7 @@ private:
   void HandleMessage(uint8_t msgClass, uint8_t msgID, const uint8_t* payload, uint16_t len);
 
   I2CBusManager& bus_manager_;
-  uint8_t address_;
+  uint16_t address_;
   mutable std::recursive_mutex mutex_;
 
   PVTData pvt_data_;
